requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
imageio==2.22.4
joblib==1.2.0
kiwisolver==1.4.4
kml2geojson==5.1.0
matplotlib==3.6.2
multidict==6.0.2
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.2
patsy==0.5.3
Pillow==9.3.0
pooch==1.6.0
pymssql==2.2.7
PyMySQL==1.0.2
pyodbc==4.0.35
pyparsing==3.0.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.6
pyvista==0.37.0
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
scooby==0.7.0
six==1.16.0
statsmodels==0.13.5
threadpoolctl==3.1.0
tqdm==4.64.1
urllib3==1.26.13
vtk==9.2.2
wslink==1.9.1
yarl==1.8.1
